requirements.txt
# AI Integration System Dependencies

# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Database and Storage
pymongo>=4.0.0
redis>=4.5.0

# HTTP and WebSocket
httpx>=0.24.0
aiohttp>=3.8.0
python-socketio>=5.8.0
websockets>=11.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
python-multipart>=0.0.6

# Logging and Monitoring
structlog>=23.0.0

# Security (for future authentication)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development and Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-json-report>=1.5.0
pytest-cov>=4.0.0

# Agent-specific dependencies (add as needed)
# numpy>=1.24.0
# pandas>=2.0.0
# scikit-learn>=1.3.0
